change return type getNumColumns() to int
add default constructor for DDF

+++ core/src/main/java/com/adatao/ddf/content/Schema.java
@@ -122,7 +122,7 @@ public class Schema implements Serializable {
    * 
    * @return number of columns
    */
-  public long getNumColumns() {
+  public int getNumColumns() {
     return this.mColumns.size();
   }
